+++ hierarchical_sales_forecasting/b2b_revenue_forecasting/__init__.py
@@ -6,8 +6,9 @@ from b2b_revenue_forecasting.quota_cascader import QuotaCascader, GateAllocation
 from b2b_revenue_forecasting.commit_reconciler import CommitReconciler
 from b2b_revenue_forecasting.pipeline_adjuster import PipelineAdjuster
 from b2b_revenue_forecasting.metric_spec import MetricSpec
+from b2b_revenue_forecasting.batch import cascade_many
 
-__version__ = "0.6.1"
+__version__ = "0.7.0"
 
 __all__ = [
     "SalesHierarchy",
@@ -17,5 +18,6 @@ __all__ = [
     "CommitReconciler",
     "PipelineAdjuster",
     "MetricSpec",
+    "cascade_many",
     "__version__",
 ]